Module espr::parser

Parser combinator for EXPRESS language

This submodule responsible for tokenize of EXPRESS language input into a SyntaxTree struct, and provides parser combinators based on nom.

Most parsers correspond to EXPRESS language grammers defined in ISO-10303-11. Each documentation of the parsers contains wirth syntax notation (WSN). For example,

124 digit = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 .

digit matches to a digit character. The head number, 124, is the serial number of this WSN in ISO-10303-11.

Remarks

EXPRESS language has two forms of “remarks” which corresponds to “comments” in Rust

(* this is called "embedded remark" *)
-- this is called "tail remark"

Each of remarks are handled by embedded_remark and tail_remark. For handling remarks appear in arbitrary position, we re-define nom’s combinators in combinator to stack remarks onto a Vec<Remark> in appeared order.

Modules

combinator

Parser combinators for EXPRESS language with remarks

Constants

KEYWORDS

Functions

abstract_entity_declaration

164 abstract_entity_declaration = ABSTRACT .

abstract_supertype

165 abstract_supertype = ABSTRACT SUPERTYPE ; .

abstract_supertype_declaration

166 abstract_supertype_declaration = ABSTRACT SUPERTYPE [ subtype_constraint ] .

actual_parameter_list

167 actual_parameter_list = ( parameter { , parameter } ) .

add_like_op

168 add_like_op = + | - | OR | XOR .

aggregate_initializer

169 aggregate_initializer = [ [ element { , element } ] ] .

aggregate_source

170 aggregate_source = simple_expression .

aggregate_type

171 aggregate_type = AGGREGATE [ : type_label ] OF parameter_type .

aggregation_types

172 aggregation_types = array_type | bag_type | list_type | set_type .

algorithm_head

173 algorithm_head = { declaration } [ constant_decl ] [ local_decl ] .

alias_stmt

174 alias_stmt = ALIAS variable_id FOR general_ref { qualifier } ; stmt { stmt } END_ALIAS ; .

array_type

175 array_type = ARRAY bound_spec OF [ OPTIONAL ] [ UNIQUE ] instantiable_type .

assignment_stmt

176 assignment_stmt = general_ref { qualifier } := expression ; .

attribute_decl

177 attribute_decl = attribute_id | redeclared_attribute .

attribute_id

attribute_qualifier

179 attribute_qualifier = . attribute_ref .

attribute_ref

bag_type

180 bag_type = BAG [ bound_spec ] OF instantiable_type .

binary_type

181 binary_type = BINARY [ width_spec ] .

boolean_type

182 boolean_type = BOOLEAN .

bound_1

183 bound_1 = numeric_expression .

bound_2

184 bound_2 = numeric_expression .

bound_spec

185 bound_spec = [ bound_1 : bound_2 ] .

built_in_constant

186 built_in_constant = CONST_E | PI | SELF | ? .

built_in_function

187 built_in_function = ABS | ACOS | ASIN | ATAN | BLENGTH | COS | EXISTS | EXP | FORMAT | HIBOUND | HIINDEX | LENGTH | LOBOUND | LOINDEX | LOG | LOG2 | LOG10 | NVL | ODD | ROLESOF | SIN | SIZEOF | SQRT | TAN | TYPEOF | USEDIN | VALUE | VALUE_IN | VALUE_UNIQUE .

built_in_procedure

188 built_in_procedure = INSERT | REMOVE .

case_action

189 case_action = case_label { , case_label } : stmt .

case_label

190 case_label = expression .

case_stmt

191 case_stmt = CASE selector OF { case_action } [ OTHERWISE : stmt ] END_CASE ; .

compound_stmt

192 compound_stmt = BEGIN stmt { stmt } END ; .

concrete_types

193 concrete_types = aggregation_types | simple_types | type_ref.

constant_body

194 constant_body = constant_id : instantiable_type := expression ; .

constant_decl

195 constant_decl = CONSTANT constant_body { constant_body } END_CONSTANT ; .

constant_factor

196 constant_factor = built_in_constant | constant_ref .

constant_id

constant_ref

constructed_types

198 constructed_types = enumeration_type | select_type .

declaration

199 declaration = entity_decl | function_decl | procedure_decl | subtype_constraint_decl | type_decl .

derive_clause

201 derive_clause = DERIVE derived_attr { derived_attr } .

derived_attr

200 derived_attr = attribute_decl : parameter_type := expression ; .

digit

124 digit = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 .

domain_rule

202 domain_rule = [ rule_label_id : ] expression .

element

203 element = expression [ : repetition ] .

embedded_remark

999 embedded_remark

encoded_character

126 encoded_character = octet octet octet octet .

encoded_string_literal

140 encoded_string_literal = " encoded_character { encoded_character } " .

entity_body

204 entity_body = { explicit_attr } [ derive_clause ] [ inverse_clause ] [ unique_clause ] [ where_clause ] .

entity_constructor

205 entity_constructor = entity_ref ( [ expression { , expression } ] ) .

entity_decl

206 entity_decl = entity_head entity_body END_ENTITY ; .

entity_head

207 entity_head = ENTITY entity_id subsuper ; .

entity_id

entity_ref

enumeration_id

enumeration_items

211 enumeration_items = ( enumeration_id { , enumeration_id } ) .

enumeration_ref

enumeration_reference

212 enumeration_reference = [ type_ref . ] enumeration_ref .

enumeration_type

213 enumeration_type = [ EXTENSIBLE ] ENUMERATION [ ( OF enumeration_items ) | enumeration_extension ] .

escape_stmt

214 escape_stmt = ESCAPE ; .

explicit_attr

215 explicit_attr = attribute_decl { , attribute_decl } : [ OPTIONAL ] parameter_type ; .

expression

216 expression = simple_expression [ rel_op_extended simple_expression ] .

factor

217 factor = simple_factor [ ** simple_factor ] .

formal_parameter

218 formal_parameter = parameter_id { , parameter_id } : parameter_type .

function_call

999 function_call = ( built_in_function | function_ref ) actual_parameter_list .

function_decl

220 function_decl = function_head algorithm_head stmt { stmt } END_FUNCTION ; .

function_head

221 function_head = FUNCTION function_id [ ( formal_parameter { ; formal_parameter } ) ] : parameter_type ; .

function_id

function_ref

general_aggregation_types

224 general_aggregation_types = general_array_type | general_bag_type | general_list_type | general_set_type .

general_array_type

225 general_array_type = ARRAY [ bound_spec ] OF [ OPTIONAL ] [ UNIQUE ] parameter_type .

general_bag_type

226 general_bag_type = BAG [ bound_spec ] OF parameter_type .

general_list_type

227 general_list_type = LIST [ bound_spec ] OF [ UNIQUE ] parameter_type .

general_ref

228 general_ref = parameter_ref | variable_ref .

general_set_type

229 general_set_type = SET [ bound_spec ] OF parameter_type .

generalized_types

223 generalized_types = aggregate_type | general_aggregation_types | generic_entity_type | generic_type .

generic_entity_type

230 generic_entity_type = GENERIC_ENTITY [ : type_label ] .

generic_type

231 generic_type = GENERIC [ : type_label ] .

group_qualifier

232 group_qualifier = \ entity_ref .

hex_digit

127 hex_digit = digit | a | b | c | d | e | f .

if_stmt

233 if_stmt = IF logical_expression THEN stmt { stmt } [ ELSE stmt { stmt } ] END_IF ; .

increment

234 increment = numeric_expression .

increment_control

235 increment_control = variable_id := bound_1 TO bound_2 [ BY increment ] .

index

236 index = numeric_expression .

index_1

237 index_1 = index .

index_2

238 index_2 = index .

index_qualifier

239 index_qualifier = [ index_1 [ : index_2 ] ] .

instantiable_type

240 instantiable_type = concrete_types | entity_ref .

integer_literal

141 integer_literal = digits .

integer_type

241 integer_type = INTEGER .

interface_specification

242 interface_specification = reference_clause | use_clause .

interval

243 interval = { interval_low interval_op interval_item interval_op interval_high } .

interval_high

244 interval_high = simple_expression .

interval_item

245 interval_item = simple_expression .

interval_low

246 interval_low = simple_expression .

interval_op

247 interval_op = < | <= .

inverse_attr

248 inverse_attr = attribute_decl : [ ( SET | BAG ) [ bound_spec ] OF ] entity_ref FOR [ entity_ref . ] attribute_ref ; .

inverse_clause

249 inverse_clause = INVERSE inverse_attr { inverse_attr } .

is_reserved

letter

128 letter = a | b | c | d | e | f | g | h | i | j | k | l |m | n | o | p | q | r | s | t | u | v | w | x |y | z .

list_type

250 list_type = LIST [ bound_spec ] OF [ UNIQUE ] instantiable_type .

literal

251 literal = binary_literal | logical_literal | real_literal | string_literal .

local_decl

252 local_decl = LOCAL local_variable { local_variable } END_LOCAL ; .

local_variable

253 local_variable = variable_id { , variable_id } : parameter_type [ := expression ] ; .

logical_expression

254 logical_expression = expression .

logical_literal

255 logical_literal = FALSE | TRUE | UNKNOWN .

logical_type

256 logical_type = LOGICAL .

multiplication_like_op

257 multiplication_like_op = * | / | DIV | MOD | AND | || .

named_type_or_rename

259 named_type_or_rename = named_types [ AS ( entity_id | type_id ) ] .

named_types

258 named_types = entity_ref | type_ref .

null_stmt

260 null_stmt = ; .

number_type

261 number_type = NUMBER .

numeric_expression

262 numeric_expression = simple_expression .

octet

136 octet = hex_digit hex_digit .

one_of

263 one_of = ONEOF ( supertype_expression { , supertype_expression } ) .

parameter

264 parameter = expression .

parameter_id

parameter_ref

parameter_type

266 parameter_type = generalized_types | named_types | simple_types .

population

267 population = entity_ref .

power_op

999 power_op = **

primary

269 primary = literal | ( qualifiable_factor { qualifier } ) .

procedure_call_stmt

270 procedure_call_stmt = ( built_in_procedure | procedure_ref ) [ actual_parameter_list ] ; .

procedure_decl

271 procedure_decl = procedure_head algorithm_head { stmt } END_PROCEDURE ; .

procedure_head

272 procedure_head = PROCEDURE procedure_id [ ( [ VAR ] formal_parameter { ; [ VAR ] formal_parameter } ) ] ; .

procedure_id

procedure_ref

qualifiable_factor

274 qualifiable_factor = attribute_ref | constant_factor | function_call | general_ref | population .

qualified_attribute

275 qualified_attribute = SELF group_qualifier attribute_qualifier .

qualifier

276 qualifier = attribute_qualifier | group_qualifier | index_qualifier .

query_expression

277 query_expression = QUERY ( variable_id <* aggregate_source | logical_expression ) .

real_literal

142 real_literal = integer_literal | ( digits . [ digits ] [ e [ sign ] digits ] ) .

real_type

278 real_type = REAL [ ( precision_spec ) ] .

redeclared_attribute

279 redeclared_attribute = qualified_attribute [ RENAMED attribute_id ] .

reference_clause

281 reference_clause = REFERENCE FROM schema_ref [ ( resource_or_rename { , resource_or_rename } ) ] ; .

referenced_attribute

280 referenced_attribute = attribute_ref | qualified_attribute .

rel_op

282 rel_op = < | > | <= | >= | <> | = | :<>: | :=: .

rel_op_extended

283 rel_op_extended = rel_op | IN | LIKE .

remark_tag

147 remark_tag = " remark_ref { . remark_ref } " .

rename_id

284 rename_id = constant_id | entity_id | function_id | procedure_id | type_id .

repeat_control

285 repeat_control = [ increment_control ] [ while_control ] [ until_control ] .

repeat_stmt

286 repeat_stmt = REPEAT repeat_control ; stmt { stmt } END_REPEAT ; .

repetition

287 repetition = numeric_expression .

resource_or_rename

288 resource_or_rename = resource_ref [ AS rename_id ] .

resource_ref

289 resource_ref = constant_ref | entity_ref | function_ref | procedure_ref | type_ref .

return_stmt

290 return_stmt = RETURN [ ( expression ) ] ; .

rule_decl

291 rule_decl = rule_head algorithm_head { stmt } where_clause END_RULE ; .

rule_head

292 rule_head = RULE rule_id FOR ( entity_ref { , entity_ref } ) ; .

rule_id

rule_label_id

rule_label_ref

rule_ref

schema_body

295 schema_body = { interface_specification } [ constant_decl ] { declaration | rule_decl } .

schema_decl

296 schema_decl = SCHEMA schema_id [ schema_version_id ] ; schema_body END_SCHEMA ; .

schema_id

schema_ref

select_extension

300 select_extension = BASED_ON type_ref [ WITH select_list ] .

select_list

301 select_list = ( named_types { , named_types } ) .

select_type

302 select_type = [ EXTENSIBLE [ GENERIC_ENTITY ] ] SELECT [ select_list | select_extension ] .

selector

299 selector = expression .

set_type

303 set_type = SET [ bound_spec ] OF instantiable_type .

simple_expression

305 simple_expression = term { add_like_op term } .

simple_factor

306 simple_factor = aggregate_initializer | entity_constructor | enumeration_reference | interval | query_expression | ( [ unary_op ] ( ( expression ) | primary ) ) .

simple_id

143 simple_id = letter { letter | digit | _ } . According to the standard, identifiers cannot be reserved keywords.

simple_string_literal

144 simple_string_literal = \q { ( \q \q ) | not_quote | \s | \x9 | \xA | \xD } \q .

simple_types

307 simple_types = binary_type | boolean_type | integer_type | logical_type | number_type | real_type | string_type .

skip_stmt

308 skip_stmt = SKIP ; .

spaces_or_remarks

Match to spaces or remarks

stmt

309 stmt = alias_stmt | assignment_stmt | case_stmt | compound_stmt | escape_stmt | if_stmt | null_stmt | procedure_call_stmt | repeat_stmt | return_stmt | skip_stmt .

string_literal

310 string_literal = simple_string_literal | encoded_string_literal .

string_type

311 string_type = STRING [ width_spec ] .

subsuper

312 subsuper = [ supertype_constraint ] [ subtype_declaration ] .

subtype_constraint

313 subtype_constraint = OF ( supertype_expression ) .

subtype_constraint_body

314 subtype_constraint_body = [ abstract_supertype ] [ total_over ] [ supertype_expression ; ] .

subtype_constraint_decl

315 subtype_constraint_decl = subtype_constraint_head subtype_constraint_body END_SUBTYPE_CONSTRAINT ; .

subtype_constraint_head

316 subtype_constraint_head = SUBTYPE_CONSTRAINT subtype_constraint_id FOR entity_ref ; .

subtype_constraint_id

subtype_constraint_ref

subtype_declaration

318 subtype_declaration = SUBTYPE OF ( entity_ref { , entity_ref } ) .

supertype_constraint

319 supertype_constraint = abstract_entity_declaration | abstract_supertype_declaration | supertype_rule .

supertype_expression

320 supertype_expression = supertype_factor { ANDOR supertype_factor } .

supertype_factor

321 supertype_factor = supertype_term { AND supertype_term } .

supertype_rule

322 supertype_rule = SUPERTYPE subtype_constraint .

supertype_term

323 supertype_term = entity_ref | one_of | ( supertype_expression ) .

tail_remark

999 tail_remark

term

325 term = factor { multiplication_like_op factor } .

total_over

326 total_over = TOTAL_OVER ( entity_ref { , entity_ref } ) ; .

type_decl

327 type_decl = TYPE type_id = underlying_type ; [ where_clause ] END_TYPE ; .

type_id

type_label

329 type_label = type_label_id | type_label_ref .

type_label_id

type_label_ref

type_ref

unary_op

331 unary_op = + | - | NOT .

underlying_type

332 underlying_type = concrete_types | constructed_types .

unique_clause

333 unique_clause = UNIQUE unique_rule ; { unique_rule ; } .

unique_rule

334 unique_rule = [ rule_label_id : ] referenced_attribute { , referenced_attribute } .

until_control

335 until_control = UNTIL logical_expression .

use_clause

336 use_clause = USE FROM schema_ref [ ( named_type_or_rename { , named_type_or_rename } ) ] ; .

variable_id

variable_ref

where_clause

338 where_clause = WHERE domain_rule ; { domain_rule ; } .

while_control

339 while_control = WHILE logical_expression .

width_spec

341 width_spec = ( width ) [ FIXED ] .